refactor: main.dart から app.dart に切り出し
